+++ pkg/local_object_storage/blobstor/fstree/getstream_test.go
@@ -8,14 +8,18 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/nspcc-dev/neofs-node/pkg/core/object"
 	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/blobstor/compression"
+	"github.com/nspcc-dev/neofs-node/pkg/util"
+	cidtest "github.com/nspcc-dev/neofs-sdk-go/container/id/test"
 	objectSDK "github.com/nspcc-dev/neofs-sdk-go/object"
+	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
 	oidtest "github.com/nspcc-dev/neofs-sdk-go/object/id/test"
 	"github.com/stretchr/testify/require"
 )
 
 func TestGetStream(t *testing.T) {
-	tree := New(WithPath(t.TempDir()))
+	tree := setupFSTree(t)
 
 	payloadSizes := []int{
 		1,
@@ -74,6 +78,66 @@ func TestGetStream(t *testing.T) {
 			})
 		}
 	})
+
+	t.Run("specific combined object", func(t *testing.T) {
+		fsTree := setupFSTree(t)
+
+		const numObjects = 3
+		const size1 = objectSDK.MaxHeaderLen - 77 - 38 - 37 // (MaxHeaderLen - headerLen - combinedDataOff) - (combinedDataOff - 1)
+		const size2 = objectSDK.MaxHeaderLen - 77 - 38 - 1  // (MaxHeaderLen - headerLen - combinedDataOff) - (combinedDataOff - 37)
+		// In this case, after the second read from the file,
+		// the third object's calculated buffer length should be sufficient
+		// to parse the combined prefix. However, if it is not expanded,
+		// its actual length will be one, which causes panic.
+		// Therefore, the buffer must have enough bytes for parsing the combined prefix
+		// only if it has been properly expanded.
+
+		cnr := cidtest.ID()
+		payload1 := make([]byte, size1)
+		_, _ = rand.Read(payload1)
+		payload2 := make([]byte, size2)
+		_, _ = rand.Read(payload2)
+
+		objects := make([]*objectSDK.Object, numObjects)
+		for i := range objects {
+			obj := objectSDK.New()
+			obj.SetID(oidtest.ID())
+			obj.SetContainerID(cnr)
+			if i == 1 {
+				obj.SetPayload(payload2)
+			} else {
+				obj.SetPayload(payload1)
+			}
+
+			objects[i] = obj
+		}
+
+		// Don't use map with FSTree.PutBatch because we need ordered writes
+		writeDataUnits := make([]writeDataUnit, 0, len(objects))
+		for _, obj := range objects {
+			addr := object.AddressOf(obj)
+			p := fsTree.treePath(addr)
+			require.NoError(t, util.MkdirAllX(filepath.Dir(p), fsTree.Permissions))
+			writeDataUnits = append(writeDataUnits, writeDataUnit{
+				id:   addr.Object(),
+				path: p,
+				data: fsTree.Compress(obj.Marshal()),
+			})
+		}
+		require.NoError(t, fsTree.writer.writeBatch(writeDataUnits))
+
+		for i := range objects {
+			res, reader, err := fsTree.GetStream(object.AddressOf(objects[i]))
+			require.NoError(t, err)
+			require.Equal(t, objects[i].CutPayload(), res)
+
+			require.NotNil(t, reader)
+			streamedPayload, err := io.ReadAll(reader)
+			require.NoError(t, err)
+			require.Equal(t, objects[i].Payload(), streamedPayload)
+			require.NoError(t, reader.Close())
+		}
+	})
 }
 
 func TestGetStreamAfterErrors(t *testing.T) {
@@ -121,4 +185,41 @@ func TestGetStreamAfterErrors(t *testing.T) {
 		_, _, err = tree.GetStream(addr)
 		require.Error(t, err)
 	})
+
+	t.Run("ID not found in combined object", func(t *testing.T) {
+		fsTree := setupFSTree(t)
+
+		cnr := cidtest.ID()
+
+		idStr := "HhW47uw6Fs48M2GtNx1Joem6qUjP1JxGPo4ffvp4QJaL"
+		id, err := oid.DecodeString(idStr)
+		require.NoError(t, err)
+		obj := objectSDK.New()
+		obj.SetID(id)
+		obj.SetContainerID(cnr)
+		addr := object.AddressOf(obj)
+
+		require.NoError(t, fsTree.Put(addr, obj.Marshal()))
+
+		p := fsTree.treePath(addr)
+
+		newIDStr := "HhW47uw6Fs48M2GtNx1Joem6qUjP1JxGPo4ffvp4QJaM"
+		newID, err := oid.DecodeString(newIDStr)
+		require.NoError(t, err)
+		newAddr := oid.NewAddress(cnr, newID)
+		newPath := fsTree.treePath(newAddr)
+
+		require.NoError(t, os.Rename(p, newPath))
+
+		_, _, err = fsTree.GetStream(newAddr)
+		require.ErrorIs(t, err, io.ErrUnexpectedEOF)
+	})
+}
+
+func setupFSTree(t *testing.T) *FSTree {
+	tree := New(WithPath(t.TempDir()))
+	require.NoError(t, tree.Open(false))
+	require.NoError(t, tree.Init())
+	t.Cleanup(func() { require.NoError(t, tree.Close()) })
+	return tree
 }