+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Changelog for NeoFS Node
 - SN no longer caches request-dependent container session errors (#3526)
 - SN no longer caches request-dependent bearer token errors (#3527)
 - Unclear SN status message for bearer token issuer mismatch (#3527)
+- Panic due to buffer size when extracting header from combined object in FSTree (#3518)
+- Endless loop when extracting header from combined object in FSTree (#3518)
 
 ### Changed
 - SN no longer accepts PUT eACL requests without container ID (#3515)